+++ android/src/main/kotlin/com/philaphonh/flutter_sunyard_s200/PrinterMethodCallHandler.kt
@@ -1,4 +1,4 @@
-package com.philaphonh.flutter_sunyard_i80
+package com.philaphonh.flutter_sunyard_s200
 
 import com.socsi.exception.SDKException
 import com.socsi.smartposapi.printer.Align