+++ example/android/app/src/main/kotlin/com/philaphonh/flutter_sunyard_s200_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.philaphonh.flutter_sunyard_i80_example
+package com.philaphonh.flutter_sunyard_s200_example
 
 import io.flutter.embedding.android.FlutterActivity