+++ example/test/widget_test.dart
@@ -8,7 +8,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:flutter_sunyard_i80_example/main.dart';
+import 'package:flutter_sunyard_s200_example/main.dart';
 
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {