+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.1.0
 
-- Rename the plugin's namespace to 'com.philaphonh.flutter_sunyard_i80'
+- Rename the plugin's namespace to 'com.philaphonh.flutter_sunyard_s200'
 - Add the namespace property to the plugin and its example project to support AGP 8 or newer
 - Upgrade AGP version for the example project to 8.1.0